rpk: move bundle code to a separate packaga
This will grow even more once we add the k8s steps
