+++ src/go/rpk/pkg/cli/cmd/debug/bundle/bundle_all.go
@@ -9,7 +9,7 @@
 
 //go:build !linux
 
-package debug
+package bundle
 
 import (
 	"context"
