+++ src/go/rpk/pkg/cli/cmd/debug/bundle/bundle_linux.go
@@ -9,7 +9,7 @@
 
 //go:build linux
 
-package debug
+package bundle
 
 import (
 	"archive/zip"
@@ -35,7 +35,6 @@ import (
 	"github.com/beevik/ntp"
 	"github.com/docker/go-units"
 	"github.com/hashicorp/go-multierror"
-	cp "github.com/otiai10/copy"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/api/admin"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/config"
 	osutil "github.com/redpanda-data/redpanda/src/go/rpk/pkg/os"
@@ -139,7 +138,7 @@ func executeBundle(ctx context.Context, bp bundleParams) error {
 		saveIP(ctx, ps),
 		saveLspci(ctx, ps),
 		saveDmidecode(ctx, ps),
-		saveControllerLogDir(bp.fs, ps, bp.cfg, bp.controllerLogLimitBytes),
+		saveControllerLogDir(ps, bp.cfg, bp.controllerLogLimitBytes),
 	}
 
 	for _, s := range steps {
@@ -221,14 +220,18 @@ func writeFileToZip(ps *stepParams, filename string, contents []byte) error {
 }
 
 // writeDirToZip walks the 'srcDir' and writes the content in 'destDir'
-// directory in the zip writer.
-func writeDirToZip(ps *stepParams, srcDir, destDir string) error {
+// directory in the zip writer. It will exclude the files that match the
+// 'exclude' regexp.
+func writeDirToZip(ps *stepParams, srcDir, destDir string, exclude *regexp.Regexp) error {
 	return filepath.WalkDir(srcDir, func(_ string, f fs.DirEntry, err error) error {
 		if err != nil {
 			return err
 		}
 		if !f.IsDir() {
 			filename := f.Name()
+			if exclude != nil && exclude.MatchString(filename) {
+				return nil
+			}
 			file, err := os.ReadFile(filepath.Join(srcDir, filename))
 			if err != nil {
 				return err
@@ -729,12 +732,12 @@ func saveDmidecode(ctx context.Context, ps *stepParams) step {
 	}
 }
 
-func saveControllerLogDir(cmdFs afero.Fs, ps *stepParams, cfg *config.Config, logLimitBytes int) step {
+func saveControllerLogDir(ps *stepParams, cfg *config.Config, logLimitBytes int) step {
 	return func() error {
 		controllerDir := filepath.Join(cfg.Redpanda.Directory, "redpanda", "controller", "0_0")
 
 		// We don't need the .base_index files to parse out the messages.
-		exclude := regexp.MustCompile(`^*.base_index`)
+		exclude := regexp.MustCompile(`^*.base_index$`)
 		size, err := osutil.DirSize(controllerDir, exclude)
 		if err != nil {
 			return err
@@ -745,26 +748,7 @@ func saveControllerLogDir(cmdFs afero.Fs, ps *stepParams, cfg *config.Config, lo
 			return fmt.Errorf("controller logs not stored, size is too big (%v). You can adjust the limit by changing --controller-logs-size-limit flag", units.HumanSize(float64(size)))
 		}
 
-		// Create a temporary directory to store the filtered files and add them
-		// to the zip writer.
-		tmpDir, err := afero.TempDir(cmdFs, "", "controller-")
-		if err != nil {
-			return err
-		}
-		defer cmdFs.RemoveAll(tmpDir)
-
-		copyOpts := cp.Options{
-			Skip: func(_ os.FileInfo, src, _ string) (bool, error) {
-				return exclude.MatchString(src), nil
-			},
-			PreserveTimes: true,
-		}
-		err = cp.Copy(controllerDir, tmpDir, copyOpts)
-		if err != nil {
-			return err
-		}
-
-		return writeDirToZip(ps, tmpDir, "controller")
+		return writeDirToZip(ps, controllerDir, "controller", exclude)
 	}
 }
