+++ src/go/rpk/pkg/cli/cmd/debug/bundle/bundle_test.go
@@ -10,7 +10,7 @@
 //go:build linux
 // +build linux
 
-package debug
+package bundle
 
 import (
 	"io"
